FEM: new member tools, move some analysis member methods from utils
