+++ src/Mod/Fem/femguiobjects/_ViewProviderFemConstraintFlowVelocity.py
@@ -34,6 +34,7 @@ from . import ViewProviderFemConstraint
 
 # for the panel
 import femtools.femutils as femutils
+import femtools.membertools as membertools
 from FreeCAD import Units
 from . import FemSelectionWidgets
 
@@ -70,7 +71,7 @@ class _TaskPanel(object):
         self._initParamWidget()
         self.form = [self._refWidget, self._paramWidget]
         analysis = obj.getParentGroup()
-        self._mesh = femutils.get_single_member(analysis, "Fem::FemMeshObject")
+        self._mesh = membertools.get_single_member(analysis, "Fem::FemMeshObject")
         self._part = None
         if self._mesh is not None:
             self._part = femutils.get_part_to_mesh(self._mesh)
