+++ src/Mod/Fem/femsolver/z88/tasks.py
@@ -32,6 +32,7 @@ import os.path
 
 import FreeCAD
 import femtools.femutils as femutils
+import femtools.membertools as membertools
 import feminout.importZ88O2Results as importZ88O2Results
 
 from .. import run
@@ -55,7 +56,7 @@ class Prepare(run.Prepare):
         w = writer.FemInputWriterZ88(
             self.analysis,
             self.solver,
-            femutils.get_mesh_to_solve(self.analysis)[0],  # pre check has been done already
+            membertools.get_mesh_to_solve(self.analysis)[0],  # pre check has been done already
             c.materials_linear,
             c.materials_nonlinear,
             c.constraints_fixed,
@@ -129,7 +130,7 @@ class Results(run.Results):
         self.load_results_z88o2()
 
     def purge_results(self):
-        for m in femutils.get_member(self.analysis, "Fem::FemResultObject"):
+        for m in membertools.get_member(self.analysis, "Fem::FemResultObject"):
             if femutils.is_of_type(m.Mesh, "Fem::FemMeshResult"):
                 self.analysis.Document.removeObject(m.Mesh.Name)
             self.analysis.Document.removeObject(m.Name)
@@ -187,6 +188,6 @@ class _Container(object):
         self.constraints_transform = []
 
     def get_several_member(self, t):
-        return femutils.get_several_member(self.analysis, t)
+        return membertools.get_several_member(self.analysis, t)
 
 ##  @}
